add existence check

+++ tests/setup_test.py
@@ -31,6 +31,7 @@ def cleanup_after_test(temp_dir):
         if Path(temp_dir, path).exists():
             shutil.rmtree(Path(temp_dir, path))
 
+
 @pytest.fixture()
 def make_folders(temp_dir):
     for path in [
@@ -50,9 +51,15 @@ def clean_up_test_installation():
     )
 
     for _, path in cfg["Directories"].items():
-        shutil.rmtree(Path(path).expanduser(), ignore_errors=True)
-    shutil.rmtree(Path(user_config_dir()) / "iSparrow_tests", )
-    shutil.rmtree(Path(user_cache_dir()) / "iSparrow_tests", )
+        if Path(path).expanduser().exists():
+            shutil.rmtree(Path(path).expanduser(), ignore_errors=True)
+    shutil.rmtree(
+        Path(user_config_dir()) / "iSparrow_tests",
+    )
+    shutil.rmtree(
+        Path(user_cache_dir()) / "iSparrow_tests",
+    )
+
 
 def test_make_directories(temp_dir, cleanup_after_test):
     base_cfg_dirs = {
